requirements.txt
kafka_python==2.0.2
fastapi==0.87.0
clickhouse-driver==0.2.5
aiohttp==3.8.3
pydantic==1.10.2
python-dotenv==0.21.0
structlog==22.3.0
flake8==5.0.4
